moving connection management from controller to connection package
